demos/scenarios/requirements.txt
gradio>=4.19.0
pyyaml>=6.0
python-dotenv>=1.0.0
tqdm>=4.65.0
openai>=1.0.0 